+++ src/main/java/org/swing/app/view/components/form/components/wrapper/DateTimeChooserWrapper.java
@@ -120,7 +120,7 @@ public class DateTimeChooserWrapper extends InputComponentWrapperBase<LocalDateT
         final int maxChildComponentHeight = availableHeight - VERTICAL_GAP;
 
         if (this.rateOfLabelFieldWidthInTotal > 0) {
-            this.labelFieldWidth = (short) (this.rateOfLabelFieldWidthInTotal * availableHeight);
+            this.labelFieldWidth = (int) (this.rateOfLabelFieldWidthInTotal * availableHeight);
         }
         this.childComponentSizeMap.put(this.labelField, new Dimension(this.labelFieldWidth, maxChildComponentHeight));
